Ocean shader test
Created a new scene and folder in assets where I'm testing out the ocean shader. Currently it works nicely, only needs some tweaking of values to get the desired results. To use the shader, apply the water material found in the ocean test folder to a plane mesh with the collider removed, the quality of the shader will depend on the number of vertices on the plane mesh, and for that reason I have included a mesh called water with LOD groups where LOD0 has the highest amount of vertices. To change the settings of the shader simply view the mesh with the shader in inspector and then check the materials details. For variable names and references open up the shader in the shader tab and there should be a list on the right side of variables and their details includes the call for each variable.