OceanShader Speed fix
Made a number of variables public
When altering speed, the higher the number the slower it goes, + and - only decides the direction of it.
Added scale variable for normal maps so we can make it look nicer, make sure x and y have the same values, the others should remain at 0.