Applied textures
the materials for pirate version of the line and frigate seem to behave strangely. created separate materials for each pirate ship so textures could be displayed along side the red overlay color. works well with brig and treasure ship, base map + color. but when attempting to do the same with line and brig you get a strong metallic sheen and cant see the base map with textures underneath.